+++ src/app/app.component.html
@@ -1,7 +1,1 @@
-<app-testnorman></app-testnorman>
-<!--
-    <app-testluna></app-testluna>
-
-<app-testmarceln></app-testmarcel>
-    <app-testluna></app-testluna>
-    -->
+<app-maincontainer></app-maincontainer>
