--- src/app/components/add-item/add-item.component.html
@@ -1,21 +1,0 @@
-<div class="card">
-    <div class="card-content">
-        <form (ngSubmit)="onSubmit()" class="col s6">
-            <div class="row">
-                <div class="input-field col s6">
-                    <input type="text" [(ngModel)]="user.displayName" name="title">
-                    <label>Title</label>
-
-                </div>
-
-                <div class="input-field col s6">
-                    <input type="text" [(ngModel)]="user.email" name="Mail">
-                    <label>Mail</label>
-
-                </div>
-                <input type="submit" value="submit" class="btn">
-            </div>
-
-        </form>
-    </div>
-</div>
