--- src/app/components/content/content.component.html
@@ -1,2 +1,0 @@
-<app-users></app-users>
-<app-add-item></app-add-item>
