+++ src/app/components/footer/footer.component.html
@@ -1,19 +1,19 @@
-<!-- Footer mitschwebend
-<footer class="page-footer" style="position: fixed; bottom: 0px; width: 100%;">
-    <div class="footer-copyright">
-        <div class="container">
-        © 2020 Copyright Text
-        <a class="grey-text text-lighten-4 right" routerLink="#!" routerLinkActive="active">Impressum</a>
-        </div>
-      </div>
-</footer>
--->
-
 <footer class="page-footer">
   <div class="footer-copyright">
     <div class="container">
-      © 2020 Copyright Text
-      <a class="grey-text text-lighten-4 right" routerLink="#!" routerLinkActive="active">Impressum</a>
+
+
+      <a style="max-width: 36px;max-height: 36px;" href="https://twitter.com/SarahforX">
+        <img style="max-width: 36px;max-height: 36px;" alt="W3Schools"
+          src="https://image.flaticon.com/icons/svg/1383/1383265.svg" width="100" height="100"></a>
+
+      <a style="max-width: 36px;max-height: 36px;margin-left: 10px;" href="https://www.instagram.com/weird_at_last_00/">
+        <img style="max-width: 36px;max-height: 36px;" alt="W3Schools"
+          src="https://image.flaticon.com/icons/svg/1383/1383263.svg" width="100" height="100"></a>
+
+
+
+      <a class="grey-text text-lighten-4 right" routerLink="Impressum" routerLinkActive="active">Impressum</a>
     </div>
   </div>
 </footer>
