+++ src/app/components/footer/footer.component.ts
@@ -9,6 +9,8 @@ export class FooterComponent implements OnInit {
 
   constructor() { }
 
+
+
   ngOnInit(): void {
   }
