+++ src/app/components/home/home.component.html
@@ -15,7 +15,7 @@
                             entworfene Stücke aus verschiedensten Genres betrachten.</p>
                     </div>
                     <div class="card-action">
-                        <a routerLink="Images" routerLinkActive="active">See the pictures</a>
+                        <a [routerLink]="['../Images']" routerLinkActive="active">See the pictures</a>
                     </div>
                 </div>
             </div>
@@ -32,7 +32,7 @@
                         </p>
                     </div>
                     <div class="card-action">
-                        <a routerLink="Videos" routerLinkActive="active">See the videos</a>
+                        <a [routerLink]="['../Videos']" routerLinkActive="active">See the videos</a>
                     </div>
                 </div>
             </div>
@@ -54,7 +54,7 @@
                             entworfene Stücke aus verschiedensten Genres betrachten.</p>
                     </div>
                     <div class="card-action">
-                        <a routerLink="Images" routerLinkActive="active">See the pictures</a>
+                        <a [routerLink]="['../Images']" routerLinkActive="active">See the pictures</a>
                     </div>
                 </div>
             </div>
@@ -71,7 +71,7 @@
                         </p>
                     </div>
                     <div class="card-action">
-                        <a routerLink="Videos" routerLinkActive="active">See the videos</a>
+                        <a [routerLink]="['../Videos']" routerLinkActive="active">See the videos</a>
                     </div>
                 </div>
             </div>
