+++ src/app/components/impressum/impressum.component.css
@@ -1,0 +1,6 @@
+.container{
+    background-color: #636e72;
+    border-radius: 18px;
+    margin-top: 3%;
+    margin-bottom: 5%;
+}
