+++ src/app/components/navbar/navbar.component.html
@@ -29,12 +29,10 @@
 
             <div class="nav-wrapper">
 
-
-
                 <!-- blaues Logo
            <a href="#" class="brand-logo center"><img src="https://image.flaticon.com/icons/png/512/2935/2935025.png" alt="Logo"></a> -->
                 <ul id="nav-mobile" class="left">
-                    <li><a routerLink="Home" routerLinkActive="active">Home</a></li>
+                    <li><a routerLink="./Home" routerLinkActive="active">Home</a></li>
                     <li><a routerLink="Images" routerLinkActive="active">Bilder</a></li>
                     <li><a routerLink="Videos" routerLinkActive="active">Videos</a></li>
                     <li><a routerLink="Texts" routerLinkActive="active">Texte</a></li>
