--- src/app/components/testluna/testluna.component.html
@@ -1,11 +1,0 @@
-<!--<app-carousel></app-carousel>
-<app-navbar></app-navbar>
-<app-cropper></app-cropper>
-<app-card></app-card>
-<app-secondgrid></app-secondgrid>
-<app-grid></app-grid>
-<app-home></app-home>
-<app-footer></app-footer>
-<app-maincontainer></app-maincontainer>-->
-<app-navbar></app-navbar>
-<app-impressum></app-impressum>
