--- src/app/components/testluna/testluna.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-testluna',
-  templateUrl: './testluna.component.html',
-  styleUrls: ['./testluna.component.css']
-})
-export class TestlunaComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
