--- src/app/components/testmarcel/testmarcel.component.html
@@ -1,1 +1,0 @@
-<p>testmarcel works!</p>
