--- src/app/components/testmarcel/testmarcel.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-testmarcel',
-  templateUrl: './testmarcel.component.html',
-  styleUrls: ['./testmarcel.component.css']
-})
-export class TestmarcelComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
