--- src/app/components/testnorman/testnorman.component.html
@@ -1,1 +1,0 @@
-<app-maincontainer></app-maincontainer>
