--- src/app/components/testnorman/testnorman.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-testnorman',
-  templateUrl: './testnorman.component.html',
-  styleUrls: ['./testnorman.component.css']
-})
-export class TestnormanComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
