--- src/app/components/users/users.component.html
@@ -1,46 +1,0 @@
-<div *ngIf="users?.length > 0;else noUsers">
-    <div class="row">
-        <div class="col s12 m6" *ngFor="let user of users">
-            <div class="card blue-grey darken-1">
-
-                <div class="card-content white-text">
-                    <span class="card-title">{{user.name}}</span>
-                    <p>{{user.mail}}</p>
-                    <div *ngIf="editState && userToEdit.uid == user.uid">
-                        <form (ngSubmit)="updateUser(user)">
-                            <div class="row">
-                                <div class="input-field col s6">
-                                    <input type="text" [(ngModel)]="user.name" name="title">
-
-                                </div>
-
-                                <div class="input-field col s6">
-                                    <input type="text" [(ngModel)]="user.mail" name="Mail">
-
-
-                                </div>
-                                <input type="submit" value="Update" class="btn orange">
-                                <button (click)="deleteItem($event,user)" class="btn red">Delete</button>
-                            </div>
-                        </form>
-                    </div>
-
-
-                </div>
-                <div class="card-action">
-                    <a (click)="editItem($event,user)">Edit</a>
-                    <a (click)="closeItem()">Minimize</a>
-
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-
-<ng-template #noUsers>
-    <hr>
-    <h5> Ther are no Users existing</h5>
-</ng-template>
-
-
-<app-add-item></app-add-item>
