--- src/app/impressum/impressum.component.css
@@ -1,6 +1,0 @@
-.container{
-    background-color: #636e72;
-    border-radius: 18px;
-    margin-top: 3%;
-    margin-bottom: 5%;
-}
